recover from shared key if it exists
