+++ protocol/xmrmaker/swap_state.go
@@ -1,10 +1,8 @@
 package xmrmaker
 
 import (
-	"bytes"
 	"context"
 	"encoding/hex"
-	"fmt"
 	"math/big"
 	"time"
 
@@ -461,93 +459,6 @@ func (s *swapState) setContract(address ethcommon.Address) error {
 func (s *swapState) setTimeouts(t0, t1 *big.Int) {
 	s.t0 = time.Unix(t0.Int64(), 0)
 	s.t1 = time.Unix(t1.Int64(), 0)
-}
-
-// checkContract checks the contract's balance and Claim/Refund keys.
-// if the balance doesn't match what we're expecting to receive, or the public keys in the contract
-// aren't what we expect, we error and abort the swap.
-func (s *swapState) checkContract(txHash ethcommon.Hash) error {
-	tx, _, err := s.ETHClient().Raw().TransactionByHash(s.ctx, txHash)
-	if err != nil {
-		return err
-	}
-
-	if tx.To() == nil || *(tx.To()) != s.contractAddr {
-		return errInvalidETHLockedTransaction
-	}
-
-	receipt, err := s.ETHClient().WaitForReceipt(s.ctx, txHash)
-	if err != nil {
-		return fmt.Errorf("failed to get receipt for New transaction: %w", err)
-	}
-
-	if receipt.Status == 0 {
-		// swap transaction reverted
-		return errLockTxReverted
-	}
-
-	// check that New log was emitted
-	if len(receipt.Logs) == 0 {
-		return errCannotFindNewLog
-	}
-
-	var event *contracts.SwapFactoryNew
-	for _, log := range receipt.Logs {
-		event, err = s.Contract().ParseNew(*log)
-		if err == nil {
-			break
-		}
-	}
-	if err != nil {
-		return errCannotFindNewLog
-	}
-
-	if !bytes.Equal(event.SwapID[:], s.contractSwapID[:]) {
-		return errUnexpectedSwapID
-	}
-
-	// check that contract was constructed with correct secp256k1 keys
-	skOurs := s.secp256k1Pub.Keccak256()
-	if !bytes.Equal(event.ClaimKey[:], skOurs[:]) {
-		return fmt.Errorf("contract claim key is not expected: got 0x%x, expected 0x%x", event.ClaimKey, skOurs)
-	}
-
-	skTheirs := s.xmrtakerSecp256K1PublicKey.Keccak256()
-	if !bytes.Equal(event.RefundKey[:], skTheirs[:]) {
-		return fmt.Errorf("contract refund key is not expected: got 0x%x, expected 0x%x", event.RefundKey, skTheirs)
-	}
-
-	// TODO: check timeouts (#161)
-
-	// check asset of created swap
-	if types.EthAsset(s.contractSwap.Asset) != types.EthAsset(event.Asset) {
-		return fmt.Errorf("swap asset is not expected: got %v, expected %v", event.Asset, s.contractSwap.Asset)
-	}
-
-	// check value of created swap
-	if s.contractSwap.Value.Cmp(event.Value) != 0 {
-		// this should never happen
-		return fmt.Errorf("swap value and event value don't match: got %v, expected %v", event.Value, s.contractSwap.Value)
-	}
-
-	receivedAmount, err := pcommon.GetEthereumAssetAmount(
-		s.ctx,
-		s.ETHClient(),
-		s.info.ReceivedAmount,
-		types.EthAsset(s.contractSwap.Asset),
-	)
-	if err != nil {
-		return err
-	}
-
-	if s.contractSwap.Value.Cmp(receivedAmount.BigInt()) != 0 {
-		return fmt.Errorf("swap value is not expected: got %v, expected %v",
-			s.contractSwap.Value,
-			receivedAmount.BigInt(),
-		)
-	}
-
-	return nil
 }
 
 // lockFunds locks XMRMaker's funds in the monero account specified by public key
